# Conflicts:
#	README.md
#	distrib/docker/Dockerfile
#	distrib/docker/ci.Dockerfile
#	docker-compose.yml

+++ internal/services/filesmanager/models.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2025, s0up and the autobrr contributors.
+// Copyright (c) 2025-2026, s0up and the autobrr contributors.
 // SPDX-License-Identifier: GPL-2.0-or-later
 
 package filesmanager
@@ -9,7 +9,7 @@ import (
 
 // CachedFile represents a cached torrent file
 type CachedFile struct {
-	ID              int
+	ID              int64
 	InstanceID      int
 	TorrentHash     string
 	FileIndex       int
